# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInterFace"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpInterface"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/refreshTokensInterface"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Swagger",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginInterFace": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "email"
          },
          "password": {
            "type": "string",
            "description": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignUpInterface": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "email"
          },
          "password": {
            "type": "string",
            "description": "password"
          },
          "name": {
            "type": "string",
            "description": "name"
          },
          "phone": {
            "type": "string",
            "description": "phone"
          },
          "birth_day": {
            "format": "date-time",
            "type": "string",
            "description": "birth_day"
          },
          "gender": {
            "type": "boolean",
            "description": "gender"
          },
          "role": {
            "type": "string",
            "description": "role"
          },
          "skill": {
            "type": "string",
            "description": "skill"
          },
          "certification": {
            "type": "string",
            "description": "certification"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "phone",
          "birth_day",
          "gender",
          "role",
          "skill",
          "certification"
        ]
      },
      "refreshTokensInterface": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}